created separate component for combined slider
